Adding headers to identity get. (#1)
Recently I worked with Salesforce on trying to get this work while impersonating a user. They explained that I would need to send the RSID cookie with the get request to get the user information of who you are impersonating. Otherwise the server returns the logout page.

Example of returned error if the RSID cookie is not sent with the identity get.
<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><script type="text/javascript">setTimeout(function(){var redirectUrl = 'https://www.salesforce.com';if(window.localStorage && window.localStorage.getItem && window.localStorage.getItem('logoutUrl')){redirectUrl = window.localStorage.getItem('logoutUrl');}window.location =  redirectUrl;},500);</script></head><body id="logoutBody" onload="if(this.bodyOnLoad) bodyOnLoad();" onunload="if(this.bodyOnUnload) bodyOnUnload();"><!-- Logout Page --></body></html>